DDP-5924 participants lookup endpoint (#734)
* DDP-5924 - implemented endpoint participants-lookup (a PRISM-like component for searching participants): they are searched in Pepper ElasticSearch storage in indices "users" and "participants_structured".
It is implemented a service ParticipantsLookupService which is an abstract layer (so, it's possible to implement different types of participants search: from DB, from other sources). Currently it is implemented ESParticipantsLookupService providing searching from ElasticSearch.
It is implemented a utility class providing more compact Elastic Search querying builder: ElasticSearchQueryBuilderUtil.
Provide escaping of special query characters in a query string.