Attempt to fix a JSON insert error
Intermittently this insert would fail on a clean database with a JSON format error. Examination of the inserted json revealed no format issues, and was well-formed JSON. Moving the JSON object to the end of the query structure seems to have resolved the issue as I have not encountered it since.