[DSM] PEPPER-745 GP Patch Access and some PatchRoute cleaning (#2527)
* PEPPER-745
I changed PatchRoute.java to allow a GP user with kit_shipping access to change kit values
I also changed the same PatchRoute class to have more logs of what kind of request is made, and to check the access based on the realm if one is provided. I also made the call to `checkUserAccessForPatch`.
In UserUtil.java, I created the checkUserAccessForPatch() which will be matching the user id in the patch request object to the one in the token. I have taken measurements to check for both email and userId.
Added a new dependency in pom file for Apache Common Lang to use the isNumeric

* PEPPER-745
created the `checkGPUserAccessForPatch` to check for kit tableAlias in UserUtil
Added tests

* PEPPER-745 fixed tests

* PEPPER-745 fixed toString, changed name of method to `checkKitShippingAccessForPatch`

* PEPPER-745 Added try-catch